requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
rasterio>=1.2.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.6.0
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0
alembic>=1.7.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
requests>=2.26.0
tqdm>=4.62.0 